Fix and unifiy src-addr and dst-addr arguments
Currently, there is a wild mix for the long option for a source or
destination address:

$ git grep -E "(link-|)(src|dst)-address" | wc -l
85
$ git grep -E "(link-|)(src|dst)-addr[^e]" | wc -l
98

And sometimes the help page does not even match the long option in the
code (e.g. icmp6 says "--[link-](src|dst)-address", but the code expects
the short form "addr".

Since the short form "addr" seems slightly more prevalent, src/dst are
already abreviated and because --link-src/dst-address is getting a bit
long, unifying everything to the short form.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>